// src/types/patient.d.ts
export interface PatientType {
  PATIENT_ID: number;
  GENDER: string;
  AGE: number;
  ADDRESS: string;
  PHONE_NUMBER: number;
  RESIDENT_REGISTRATION_NUMBER: string;
  SPECIAL_NOTE: string;
  PHYSICAL_INFO: number;
  USER_NAME: string;
  DEPARTMENT: string;
  LAST_TREATMENT_DATE: string;
  RESERVATION_DATE: string;
  NAME: string;
}